moved App to the root, next to index.tsx
